+++ src/Additions/GUA_Freemius.php
@@ -10,6 +10,8 @@
 
 namespace Fragen\Git_Updater\Additions;
 
+use Fragen\Git_Updater\Shim;
+
 /*
  * Exit if called directly.
  */
@@ -42,7 +44,7 @@ class GUA_Freemius {
 				if ( ! isset( $gua_fs ) ) {
 
 					// Init Freemius SDK.
-					require_once dirname( __DIR__, 2 ) . '/vendor/freemius/wordpress-sdk/start.php';
+					require_once Shim::dirname( __DIR__, 2 ) . '/vendor/freemius/wordpress-sdk/start.php';
 
 					$gua_fs = fs_dynamic_init(
 						[
@@ -89,6 +91,6 @@ class GUA_Freemius {
 	 * @return string
 	 */
 	public function add_icon() {
-		return dirname( __DIR__, 2 ) . '/assets/icon.svg';
+		return Shim::dirname( __DIR__, 2 ) . '/assets/icon.svg';
 	}
 }